requirements.txt
Flask==2.3.3
pandas==2.0.3
mutagen==1.46.0
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
